+++ ManagedCorDebug/Managed/ResultTypes/ReadVirtualResult.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Diagnostics;
 
 namespace ManagedCorDebug
@@ -11,14 +12,14 @@ namespace ManagedCorDebug
         /// <summary>
         /// A pointer to a buffer that receives the data.
         /// </summary>
-        public byte buffer { get; }
+        public IntPtr buffer { get; }
 
         /// <summary>
         /// A pointer to the number of bytes returned.
         /// </summary>
         public int bytesRead { get; }
 
-        public ReadVirtualResult(byte buffer, int bytesRead)
+        public ReadVirtualResult(IntPtr buffer, int bytesRead)
         {
             this.buffer = buffer;
             this.bytesRead = bytesRead;
